Renamed Microsoft.Dnx.TestAdapter to Microsoft.Dnx.Testing.Abstractions
