+++ Testing.sln
@@ -13,8 +13,6 @@ Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Sample.Tests", "test\Sample
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.TestHost", "src\Microsoft.Dnx.TestHost\Microsoft.Dnx.TestHost.xproj", "{F003F228-2AE2-4E9D-877B-93EB773B5061}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.TestAdapter", "src\Microsoft.Dnx.TestAdapter\Microsoft.Dnx.TestAdapter.xproj", "{71C2F8AA-05E6-47DB-9A1A-D2760CEF7DC1}"
-EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.TestHost.Tests", "test\Microsoft.Dnx.TestHost.Tests\Microsoft.Dnx.TestHost.Tests.xproj", "{25D18D0C-119C-4AB4-BCA6-AC06179335FA}"
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Framework.Logging.Testing", "src\Microsoft.Framework.Logging.Testing\Microsoft.Framework.Logging.Testing.xproj", "{379AA56B-E1A6-4133-9A45-7F70385F39FB}"
@@ -38,6 +36,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.AspNet.Testing.Tests", "test\Microsoft.AspNet.Testing.Tests\Microsoft.AspNet.Testing.Tests.xproj", "{C9F758D6-17A3-4BF1-BCCD-972B1DA959FC}"
 EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Microsoft.Dnx.Testing.Abstractions", "src\Microsoft.Dnx.Testing.Abstractions\Microsoft.Dnx.Testing.Abstractions.xproj", "{DCDFE282-03DE-4DBC-B90C-CC3CE3EC8162}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -56,10 +56,6 @@ Global
 		{F003F228-2AE2-4E9D-877B-93EB773B5061}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F003F228-2AE2-4E9D-877B-93EB773B5061}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F003F228-2AE2-4E9D-877B-93EB773B5061}.Release|Any CPU.Build.0 = Release|Any CPU
-		{71C2F8AA-05E6-47DB-9A1A-D2760CEF7DC1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{71C2F8AA-05E6-47DB-9A1A-D2760CEF7DC1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{71C2F8AA-05E6-47DB-9A1A-D2760CEF7DC1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{71C2F8AA-05E6-47DB-9A1A-D2760CEF7DC1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{25D18D0C-119C-4AB4-BCA6-AC06179335FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{25D18D0C-119C-4AB4-BCA6-AC06179335FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{25D18D0C-119C-4AB4-BCA6-AC06179335FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -92,6 +88,10 @@ Global
 		{C9F758D6-17A3-4BF1-BCCD-972B1DA959FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C9F758D6-17A3-4BF1-BCCD-972B1DA959FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C9F758D6-17A3-4BF1-BCCD-972B1DA959FC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DCDFE282-03DE-4DBC-B90C-CC3CE3EC8162}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DCDFE282-03DE-4DBC-B90C-CC3CE3EC8162}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DCDFE282-03DE-4DBC-B90C-CC3CE3EC8162}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DCDFE282-03DE-4DBC-B90C-CC3CE3EC8162}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -100,7 +100,6 @@ Global
 		{09BE5203-8042-4338-9713-E44169342E72} = {5525B6EA-8BBB-4437-BD09-419AE380BBA8}
 		{25D18D0B-119C-4AB4-BCA6-AC06179335FA} = {09F799F3-E521-466F-B155-B89E2746C8C9}
 		{F003F228-2AE2-4E9D-877B-93EB773B5061} = {5525B6EA-8BBB-4437-BD09-419AE380BBA8}
-		{71C2F8AA-05E6-47DB-9A1A-D2760CEF7DC1} = {5525B6EA-8BBB-4437-BD09-419AE380BBA8}
 		{25D18D0C-119C-4AB4-BCA6-AC06179335FA} = {09F799F3-E521-466F-B155-B89E2746C8C9}
 		{379AA56B-E1A6-4133-9A45-7F70385F39FB} = {5525B6EA-8BBB-4437-BD09-419AE380BBA8}
 		{6FA3EFAE-2DBC-4532-A5A0-C2EBD8DD672F} = {09F799F3-E521-466F-B155-B89E2746C8C9}
@@ -109,5 +108,6 @@ Global
 		{90C1A9CE-E402-4206-8194-03A844435FC3} = {5525B6EA-8BBB-4437-BD09-419AE380BBA8}
 		{2A16C1BE-8777-473A-A581-02E10D82C49D} = {57A90243-F8A5-4115-BA05-A1CEEBD8A5CD}
 		{C9F758D6-17A3-4BF1-BCCD-972B1DA959FC} = {09F799F3-E521-466F-B155-B89E2746C8C9}
+		{DCDFE282-03DE-4DBC-B90C-CC3CE3EC8162} = {5525B6EA-8BBB-4437-BD09-419AE380BBA8}
 	EndGlobalSection
 EndGlobal
