+++ global.json
@@ -1,3 +1,3 @@
 {
-    "projects": ["src"]
+    "projects": ["src", "c:/github/aspnet.xunit/src"]
 }
