+++ src/Microsoft.Dnx.TestHost/project.json
@@ -17,7 +17,7 @@
             "version": "1.0.0-*",
             "type": "build"
         },
-        "Microsoft.Dnx.TestAdapter": "1.0.0-*",
+        "Microsoft.Dnx.Testing.Abstractions": "1.0.0-*",
         "Microsoft.Framework.CommandLineUtils.Sources": {
             "version": "1.0.0-*",
             "type": "build"
