+++ src/Microsoft.Dnx.Testing.Abstractions/ITestDiscoverySink.cs
@@ -1,12 +1,10 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Dnx.TestAdapter
+namespace Microsoft.Dnx.Testing.Abstractions
 {
-    public interface ITestExecutionSink
+    public interface ITestDiscoverySink
     {
-        void RecordStart(Test test);
-
-        void RecordResult(TestResult testResult);
+        void SendTest(Test test);
     }
 }
