+++ src/Microsoft.Dnx.Testing.Abstractions/ITestExecutionSink.cs
@@ -1,14 +1,12 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Dnx.TestAdapter
+namespace Microsoft.Dnx.Testing.Abstractions
 {
-    public enum TestOutcome
+    public interface ITestExecutionSink
     {
-        None,
-        Passed,
-        Failed,
-        Skipped,
-        NotFound
+        void RecordStart(Test test);
+
+        void RecordResult(TestResult testResult);
     }
 }
