+++ src/Microsoft.Dnx.Testing.Abstractions/Properties/AssemblyInfo.cs
