+++ src/Microsoft.Dnx.Testing.Abstractions/SourceInformation.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Dnx.TestAdapter
+namespace Microsoft.Dnx.Testing.Abstractions
 {
     public class SourceInformation
     {
