+++ src/Microsoft.Dnx.Testing.Abstractions/TestOutcome.cs
@@ -1,10 +1,14 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Dnx.TestAdapter
+namespace Microsoft.Dnx.Testing.Abstractions
 {
-    public interface ITestDiscoverySink
+    public enum TestOutcome
     {
-        void SendTest(Test test);
+        None,
+        Passed,
+        Failed,
+        Skipped,
+        NotFound
     }
 }
