+++ src/Microsoft.Dnx.Testing.Abstractions/project.json
